+++ includes/class-devtb-file-handler.php
@@ -6,7 +6,7 @@
  *
  * @package    DevelopmentTranslation_Bridge
  * @subpackage CLI
- * @version    4.9.0
+ * @version    4.10.0
  */
 
 class DEVTB_File_Handler {